add metadata key to socket, make common telemetry helper
