+++ lib/slipstream/connection/telemetry.ex
@@ -6,26 +6,17 @@ defmodule Slipstream.Connection.Telemetry do
 
   # helper functions for emitting telemetry about a connection
 
-  def id(length \\ 16) do
-    length
-    |> :crypto.strong_rand_bytes()
-    |> Base.encode16()
-    |> binary_part(0, length)
-    |> String.downcase()
-  end
-
-  def trace_id, do: id(32)
-
   @doc """
   Wraps the connection pipeline in order to emit telemetry for each message
   sent to the connection process
   """
+  @doc since: "0.3.0"
   def span(initial_pipeline, func) do
     metadata = %{
       start_time: DateTime.utc_now(),
       raw_message: initial_pipeline.raw_message,
       start_state: initial_pipeline.state,
-      span_id: id(),
+      span_id: Slipstream.TelemetryHelper.id(),
       connection_id: initial_pipeline.state.connection_id,
       trace_id: initial_pipeline.state.trace_id
     }
@@ -57,6 +48,7 @@ defmodule Slipstream.Connection.Telemetry do
   @doc """
   Emits the start event for a connection
   """
+  @doc since: "0.3.0"
   def begin(%State{} = state) do
     metadata = %{
       start_time: DateTime.utc_now(),
@@ -78,6 +70,7 @@ defmodule Slipstream.Connection.Telemetry do
   @doc """
   Emits the stop event for a connection
   """
+  @doc since: "0.3.0"
   def conclude(%State{} = state, reason) do
     metadata =
       state.metadata
